minor changes for consistency and readability
First I fixed a typo (it's -> its) but then I started making the list of operations more consistent. 

I removed "ontology term matching" after first changing it to "find matching terms" and concluding that was basically the same as "search an ontology for a term", but maybe there's something I'm missing.