Adding an OboGraph backend.
This also adds a new package converters, for handling
inter-data model conversion. So far this is only
OboGraph->OWL/RDF

(this goes the rdflib route rather than funowl)

Note that this means it is now possible to do:

    runoak -i obograph:foo.json -o foo.owl -O owl

Fixes #302

This PR also includes fleshing out documentation,
and the shared test suite.