Creating text definition validation framework.
Noyte this is currently turned off by default when
running validate on the command line - it may be slow
for some ontologies, the default text-annotator uses a
naive lookup strategy.

Creating a new package oaklib.validation for this,
and moving the existing linter here.

This PR also includes some minor additions to the previous
CLI docs.