Use the gzipped file from the semsql s3 cache.
This will ensure faster downloads for large databases;
it also means we can start phasing out the direct .db deposits
and save s3 space.

Utilizes https://github.com/cthoyt/pystow/issues/45

Note that the gzipped file is not cached in ~/.data - but
the gunzipped file is. It is necessary for the user
to periodically clean this cache.
See https://github.com/cthoyt/pystow/issues/54