Skip tests requiring optional modules when they are absent.
The `gilda` and `semsimian` modules are optional dependencies, but the
tests that depend do not expect that they may not be present and error
out if they cannot be imported.

Since they are optional, their absence should not cause the test suite
to fail, so this commit ensures that the corresponding tests are
properly skipped if the optional modules are not found.

closes #794