Add cache management features.
We add a cache management layer on top of Pystow. This takes the form of
two classes (both in `oaklib.utilities.caching`):

* one representing the cache management policy, i.e. the logic dictating
  whether a cached file (if present) should be refreshed or not;
* one representing the file cache itself.

The policy is set once by the main entry point method, using either a
default policy of refreshing cached data after 7 days, or another policy
explicitly selected by the user with the new `--caching` option.

The class that represents the file cache is the one that the rest of OAK
should interact with whenever an access to caching data is needed.
Ultimately, all calls to the Pystow module should be replaced to calls
to FileCache, the use of Pystow becoming an implementation detail
entirely encapsulated in FileCache.