Re-implement cache-ls and cache-clear.
Add new methods to the FileCache class to (1) get the list of files
present in the cache and (2) delete files in the cache.

Replace the implementations of the cache-ls and cache-clear commands to
use the new methods, so that the details of cache listing and clearing
remain encapsulated in FileCache.

As a side-effect, this automatically fixes the issue that cache listing
was only working on Unix-like systems, since the FileCache
implementation is pure Python and does not rely on the ls(1) Unix
command.