Implement the cache reset policy.
The intended difference between the REFRESH and RESET caching policies
is that, when a cache lookup is attempted, REFRESH should cause the file
that was looked up -- and only that file -- to be refreshed, leaving any
other file that may be present in the cache untouched. RESET, on the
other hand, should entirely clear the cache, so that not only the file
that was looked up should be refreshed, but any other file that may
looked up in a subsequent call should be refreshed as well.

This commit implements the intended behaviour for the RESET policy.