Misc documentation fix.
The "user_config_dir" returned by the Appdirs package under macOS is not
in "~/Library/Prefences" but under "~/Library/Application Support"
(Appdirs documentation is not up to date).

Also, there is no need to mention the roaming directory under Windows,
as Appdirs will never use that directory unless we explicitly asks it do
so (which we don't).

There is also no need for a show_default=True parameter with the
--caching option, since that option has _no_ default.